Fix compile warnings, add multiprocessor compilation

+++ src/gui/command/cmdConnectWire.cpp
@@ -116,7 +116,7 @@ void cmdConnectWire::sendMessagesToConnect(GUICircuit *gCircuit, IDType wireId,
 		internalHotspots.push_back(hotspot);
 	}
 	else {
-		for (int i = 0; i < wireIds.size(); i++) {
+		for (int i = 0; i < (int)wireIds.size(); i++) {
 			internalHotspots.push_back(hotspot + "_" + std::to_string(i));
 		}
 	}
@@ -124,7 +124,7 @@ void cmdConnectWire::sendMessagesToConnect(GUICircuit *gCircuit, IDType wireId,
 	bool isInput = gate->isConnectionInput(hotspot);
 
 	// Connect each of wire's bus-lines to its corresponding gate hotspot.
-	for (int i = 0; i < internalHotspots.size(); i++) {
+	for (int i = 0; i < (int)internalHotspots.size(); i++) {
 		if (isInput) {
 			gCircuit->sendMessageToCore(klsMessage::Message(klsMessage::MT_SET_GATE_INPUT,
 				new klsMessage::Message_SET_GATE_INPUT(gateId, internalHotspots[i], wireIds[i])));
@@ -158,7 +158,7 @@ void cmdConnectWire::sendMessagesToDisconnect(GUICircuit *gCircuit,
 		internalHotspots.push_back(hotspot);
 	}
 	else {
-		for (int i = 0; i < wireIds.size(); i++) {
+		for (int i = 0; i < (int)wireIds.size(); i++) {
 			internalHotspots.push_back(hotspot + "_" + to_string(i));
 		}
 	}
@@ -166,7 +166,7 @@ void cmdConnectWire::sendMessagesToDisconnect(GUICircuit *gCircuit,
 	bool isInput = gate->isConnectionInput(hotspot);
 
 	// Disconnect each of wire's bus-lines from its corresponding gate hotspot.
-	for (int i = 0; i < internalHotspots.size(); i++) {
+	for (int i = 0; i < (int)internalHotspots.size(); i++) {
 		if (isInput) {
 			gCircuit->sendMessageToCore(klsMessage::Message(klsMessage::MT_SET_GATE_INPUT,
 				new klsMessage::Message_SET_GATE_INPUT(gateId, internalHotspots[i], 0, true)));
